[INTERNAL] create-api-index.js separation from cli
Moved index creation script into separate file such that the
module can be used without command line.
